chore: refactor scrapers to interface, reorganize directories, and update tests
